commented out the Matrix.Transpose property
This was a source of frustration. After adding performant operations where you can re-use matrices instead of returning a new one the math was wrong. Transpose is evaluated only once and thereby stored. Afterwards it merely recalles the stored value what is bad if you have 're-used' your matrix in the meantime.